+++ common/gcmelee.lua
@@ -97,20 +97,22 @@ function gcmelee.DoDefault()
 
     gcinclude.DoDefaultIdle()
 
-    if (player.MainJob == 'PLD' or player.MainJob == 'NIN' or player.MainJob == 'DRK' or gcdisplay.GetToggle('Hate')) then
-        if (player.SubJob == 'NIN' or player.MainJob == 'NIN') then
-            local function GetShadowCount()
-                for buffId, shadowCount in pairs(utsuBuffs) do
-                    if (gData.GetBuffCount(buffId) > 0) then
-                        return shadowCount
+    if (not gcinclude.horizon_safe_mode) then
+        if (player.MainJob == 'PLD' or player.MainJob == 'NIN' or player.MainJob == 'DRK') then
+            if (player.SubJob == 'NIN' or player.MainJob == 'NIN') then
+                local function GetShadowCount()
+                    for buffId, shadowCount in pairs(utsuBuffs) do
+                        if (gData.GetBuffCount(buffId) > 0) then
+                            return shadowCount
+                        end
                     end
-                end
 
-                return 0
-            end
-            if (GetShadowCount() == 0) then
-                gFunc.EquipSet('IdleDT')
-                if (gcdisplay.IdleSet == 'Alternate') then gFunc.EquipSet('IdleALTDT') end
+                    return 0
+                end
+                if (GetShadowCount() == 0) then
+                    gFunc.EquipSet('IdleDT')
+                    if (gcdisplay.IdleSet == 'Alternate') then gFunc.EquipSet('IdleALTDT') end
+                end
             end
         end
     end